Fix up careless macro invocations
This resolves paste as an export from archetype itself, but hides it
from the docs; it's not part of our actual public interface, it is
only incidental in getting the test macro to work.

Also, the `env!` call would point to the wrong location, and instead
needed to be `std::env:var` which is trusted to always exist from cargo,
hence the unwrap.